pyproject.toml
[project]
name = "lbg-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.79.0",
    "httpx>=0.28.1",
    "hypercorn>=0.18.0",
    "openviking>=0.1.17",
    "python-dotenv>=1.2.1",
    "restate-sdk>=0.14.2",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
